make voxels optional
